+++ src/store/table_ops.rs
@@ -172,7 +172,7 @@ impl<'a> TableOperations<'a> {
 		let after_count = table.count_rows(None).await?;
 		let deleted_count = before_count.saturating_sub(after_count);
 
-		Ok(deleted_count as usize)
+		Ok(deleted_count)
 	}
 
 	/// Remove blocks by hashes from a table
